feat: Make all key labels the same size
Size placeholding for dynamic jump targets wasn't working.
If the size for a key label depends on the shortcut it is showing,
I cannot calculate the size of the label before the shortcut
is allocated.
To solve this, all key labels are now the size of a WW label.
This also looks cleaner when multiple labels are shown in a column.